# SendActivities Performance Optimizations