--- a/libraries/Microsoft.Bot.Builder.Core/Adapters/TestAdapter.cs
+++ b/libraries/Microsoft.Bot.Builder.Core/Adapters/TestAdapter.cs
@@ -72,17 +72,41 @@ namespace Microsoft.Bot.Builder.Adapters
 
         public async override Task<ResourceResponse[]> SendActivities(ITurnContext context, Activity[] activities)
         {
+            if (context == null)
+            {
+                throw new ArgumentNullException(nameof(context));
+            }
+
+            if (activities == null)
+            {
+                throw new ArgumentNullException(nameof(activities));
+            }
+
+            if (activities.Length == 0)
+            {
+                throw new ArgumentException("Expecting one or more activities, but the array was empty.", nameof(activities));
+            }
+
             var responses = new ResourceResponse[activities.Length];
 
-            for(var index = 0; index < activities.Length; index++)
+            /* 
+             * NOTE: we're using for here (vs. foreach) because we want to simultaneously index into the
+             * activities array to get the activity to process as well as use that index to assign
+             * the response to the responses array and this is the most cost effective way to do that.
+             */
+            for (var index = 0; index < activities.Length; index++)
             {
                 var activity = activities[index];
 
                 if (String.IsNullOrEmpty(activity.Id))
+                {
                     activity.Id = Guid.NewGuid().ToString("n");
+                }
 
                 if (activity.Timestamp == null)
+                {
                     activity.Timestamp = DateTime.UtcNow;
+                }
 
 
                 if (activity.Type == ActivityTypesEx.Delay)
@@ -92,6 +116,7 @@ namespace Microsoft.Bot.Builder.Adapters
                     // to keep the behavior as close as possible to facillitate
                     // more realistic tests.                     
                     int delayMs = (int)activity.Value;
+
                     await Task.Delay(delayMs);
                 }
                 else