Optimizations focusing _mainly_ on the `TurnContext::SendActivities` implementation, but includes a couple small optimizations for the various `BotAdapter` subclasses `SendActivities` implementations as well.

Here are the benchmark results that focus only on the `TurnContext::SendActivities` implementation optimizations:

``` ini

BenchmarkDotNet=v0.10.14, OS=Windows 10.0.17134
Intel Core i7-8650U CPU 1.90GHz (Kaby Lake R), 1 CPU, 8 logical and 4 physical cores
  [Host]     : .NET Framework 4.7.1 (CLR 4.0.30319.42000), 32bit LegacyJIT-v4.7.3101.0
  DefaultJob : .NET Framework 4.7.1 (CLR 4.0.30319.42000), 32bit LegacyJIT-v4.7.3101.0


```
|                   Method | NumberOfActivities | NumberOfHandlers |      Mean |     Error |    StdDev |    Median | Scaled | ScaledSD |  Gen 0 | Allocated |
|------------------------- |------------------- |----------------- |----------:|----------:|----------:|----------:|-------:|---------:|-------:|----------:|
|  **SendActivities_ORIGINAL** |                  **1** |                **0** |  **1.360 us** | **0.0271 us** | **0.0430 us** |  **1.351 us** |   **1.00** |     **0.00** | **0.1411** |     **596 B** |
| SendActivities_OPTIMIZED |                  1 |                0 |  1.070 us | 0.0264 us | 0.0259 us |  1.069 us |   0.79 |     0.03 | 0.0973 |     412 B |
|                          |                    |                  |           |           |           |           |        |          |        |           |
|  **SendActivities_ORIGINAL** |                  **1** |                **1** |  **1.788 us** | **0.0345 us** | **0.0323 us** |  **1.790 us** |   **1.00** |     **0.00** | **0.1907** |     **804 B** |
| SendActivities_OPTIMIZED |                  1 |                1 |  1.116 us | 0.0216 us | 0.0191 us |  1.119 us |   0.62 |     0.02 | 0.1125 |     476 B |
|                          |                    |                  |           |           |           |           |        |          |        |           |
|  **SendActivities_ORIGINAL** |                  **1** |                **2** |  **2.427 us** | **0.0488 us** | **0.1303 us** |  **2.419 us** |   **1.00** |     **0.00** | **0.2632** |    **1116 B** |
| SendActivities_OPTIMIZED |                  1 |                2 |  1.211 us | 0.0235 us | 0.0322 us |  1.205 us |   0.50 |     0.03 | 0.1240 |     524 B |
|                          |                    |                  |           |           |           |           |        |          |        |           |
|  **SendActivities_ORIGINAL** |                  **1** |                **5** |  **4.854 us** | **0.0688 us** | **0.0610 us** |  **4.841 us** |   **1.00** |     **0.00** | **0.6027** |    **2532 B** |
| SendActivities_OPTIMIZED |                  1 |                5 |  1.194 us | 0.0221 us | 0.0206 us |  1.198 us |   0.25 |     0.01 | 0.1583 |     668 B |
|                          |                    |                  |           |           |           |           |        |          |        |           |
|  **SendActivities_ORIGINAL** |                  **1** |               **10** | **16.425 us** | **0.8674 us** | **2.5575 us** | **14.906 us** |   **1.00** |     **0.00** | **1.5411** |    **6492 B** |
| SendActivities_OPTIMIZED |                  1 |               10 |  1.262 us | 0.0245 us | 0.0272 us |  1.264 us |   0.08 |     0.01 | 0.2155 |     908 B |
|                          |                    |                  |           |           |           |           |        |          |        |           |
|  **SendActivities_ORIGINAL** |                  **2** |                **0** |  **1.842 us** | **0.0350 us** | **0.0327 us** |  **1.838 us** |   **1.00** |     **0.00** | **0.1640** |     **704 B** |
| SendActivities_OPTIMIZED |                  2 |                0 |  1.571 us | 0.0305 us | 0.0286 us |  1.564 us |   0.85 |     0.02 | 0.1221 |     516 B |
|                          |                    |                  |           |           |           |           |        |          |        |           |
|  **SendActivities_ORIGINAL** |                  **2** |                **1** |  **2.280 us** | **0.0423 us** | **0.0396 us** |  **2.263 us** |   **1.00** |     **0.00** | **0.2136** |     **912 B** |
| SendActivities_OPTIMIZED |                  2 |                1 |  1.640 us | 0.0404 us | 0.0358 us |  1.627 us |   0.72 |     0.02 | 0.1373 |     580 B |
|                          |                    |                  |           |           |           |           |        |          |        |           |
|  **SendActivities_ORIGINAL** |                  **2** |                **2** |  **2.781 us** | **0.0516 us** | **0.0458 us** |  **2.765 us** |   **1.00** |     **0.00** | **0.2899** |    **1224 B** |
| SendActivities_OPTIMIZED |                  2 |                2 |  1.985 us | 0.0392 us | 0.0367 us |  1.990 us |   0.71 |     0.02 | 0.1488 |     628 B |
|                          |                    |                  |           |           |           |           |        |          |        |           |
|  **SendActivities_ORIGINAL** |                  **2** |                **5** |  **8.236 us** | **0.3242 us** | **0.9561 us** |  **8.364 us** |   **1.00** |     **0.00** | **0.6256** |    **2640 B** |
| SendActivities_OPTIMIZED |                  2 |                5 |  2.751 us | 0.0985 us | 0.2904 us |  2.792 us |   0.34 |     0.05 | 0.1831 |     772 B |
|                          |                    |                  |           |           |           |           |        |          |        |           |
|  **SendActivities_ORIGINAL** |                  **2** |               **10** | **21.072 us** | **0.4485 us** | **0.8533 us** | **20.963 us** |   **1.00** |     **0.00** | **1.5564** |    **6600 B** |
| SendActivities_OPTIMIZED |                  2 |               10 |  2.383 us | 0.0467 us | 0.0437 us |  2.385 us |   0.11 |     0.00 | 0.2403 |    1012 B |
|                          |                    |                  |           |           |           |           |        |          |        |           |
|  **SendActivities_ORIGINAL** |                  **5** |                **0** |  **4.698 us** | **0.0889 us** | **0.0873 us** |  **4.683 us** |   **1.00** |     **0.00** | **0.2518** |    **1072 B** |
| SendActivities_OPTIMIZED |                  5 |                0 |  3.757 us | 0.1190 us | 0.3434 us |  3.588 us |   0.80 |     0.07 | 0.2060 |     872 B |
|                          |                    |                  |           |           |           |           |        |          |        |           |
|  **SendActivities_ORIGINAL** |                  **5** |                **1** |  **4.246 us** | **0.0836 us** | **0.0782 us** |  **4.226 us** |   **1.00** |     **0.00** | **0.2975** |    **1280 B** |
| SendActivities_OPTIMIZED |                  5 |                1 |  3.606 us | 0.0717 us | 0.0907 us |  3.596 us |   0.85 |     0.03 | 0.2213 |     936 B |
|                          |                    |                  |           |           |           |           |        |          |        |           |
|  **SendActivities_ORIGINAL** |                  **5** |                **2** |  **4.773 us** | **0.1168 us** | **0.1092 us** |  **4.737 us** |   **1.00** |     **0.00** | **0.3738** |    **1592 B** |
| SendActivities_OPTIMIZED |                  5 |                2 |  3.624 us | 0.0700 us | 0.0687 us |  3.603 us |   0.76 |     0.02 | 0.2327 |     984 B |
|                          |                    |                  |           |           |           |           |        |          |        |           |
|  **SendActivities_ORIGINAL** |                  **5** |                **5** |  **7.470 us** | **0.1599 us** | **0.1496 us** |  **7.440 us** |   **1.00** |     **0.00** | **0.7019** |    **3008 B** |
| SendActivities_OPTIMIZED |                  5 |                5 |  3.612 us | 0.0459 us | 0.0407 us |  3.620 us |   0.48 |     0.01 | 0.2670 |    1128 B |
|                          |                    |                  |           |           |           |           |        |          |        |           |
|  **SendActivities_ORIGINAL** |                  **5** |               **10** | **17.930 us** | **0.3580 us** | **0.7473 us** | **17.792 us** |   **1.00** |     **0.00** | **1.6479** |    **6968 B** |
| SendActivities_OPTIMIZED |                  5 |               10 |  3.509 us | 0.0675 us | 0.0631 us |  3.506 us |   0.20 |     0.01 | 0.3242 |    1368 B |
|                          |                    |                  |           |           |           |           |        |          |        |           |
|  **SendActivities_ORIGINAL** |                 **10** |                **0** |  **6.193 us** | **0.1544 us** | **0.1444 us** |  **6.156 us** |   **1.00** |     **0.00** | **0.3967** |    **1688 B** |
| SendActivities_OPTIMIZED |                 10 |                0 |  6.009 us | 0.1158 us | 0.1239 us |  5.971 us |   0.97 |     0.03 | 0.3433 |    1468 B |
|                          |                    |                  |           |           |           |           |        |          |        |           |
|  **SendActivities_ORIGINAL** |                 **10** |                **1** |  **6.827 us** | **0.1362 us** | **0.1337 us** |  **6.837 us** |   **1.00** |     **0.00** | **0.4501** |    **1896 B** |
| SendActivities_OPTIMIZED |                 10 |                1 |  5.928 us | 0.0840 us | 0.0744 us |  5.907 us |   0.87 |     0.02 | 0.3586 |    1532 B |
|                          |                    |                  |           |           |           |           |        |          |        |           |
|  **SendActivities_ORIGINAL** |                 **10** |                **2** |  **7.195 us** | **0.1264 us** | **0.1121 us** |  **7.150 us** |   **1.00** |     **0.00** | **0.5188** |    **2208 B** |
| SendActivities_OPTIMIZED |                 10 |                2 |  5.928 us | 0.1114 us | 0.1042 us |  5.918 us |   0.82 |     0.02 | 0.3738 |    1580 B |
|                          |                    |                  |           |           |           |           |        |          |        |           |
|  **SendActivities_ORIGINAL** |                 **10** |                **5** |  **9.695 us** | **0.1908 us** | **0.2343 us** |  **9.694 us** |   **1.00** |     **0.00** | **0.8545** |    **3624 B** |
| SendActivities_OPTIMIZED |                 10 |                5 |  6.214 us | 0.1238 us | 0.2557 us |  6.104 us |   0.64 |     0.03 | 0.4044 |    1724 B |
|                          |                    |                  |           |           |           |           |        |          |        |           |
|  **SendActivities_ORIGINAL** |                 **10** |               **10** | **19.661 us** | **0.1826 us** | **0.1708 us** | **19.642 us** |   **1.00** |     **0.00** | **1.8005** |    **7584 B** |
| SendActivities_OPTIMIZED |                 10 |               10 |  6.297 us | 0.0471 us | 0.0417 us |  6.300 us |   0.32 |     0.00 | 0.4654 |    1964 B |